Pow / Root n
Se agrega raíz con base n y exponencial n

+++ bdd.workshop.calculator.tests.tdd/steps/Calculator.cs
@@ -19,13 +19,13 @@ namespace bdd.workshop.calculator.tests.tdd.steps
             _scenarioContext = scenarioContext;
         }
         [Given(@"the first number is (.*)")]
-        public void GivenTheFirstNumberIs(int firstNumber)
+        public void GivenTheFirstNumberIs(double firstNumber)
         {
             _scenarioContext.Add("FirstNumber", firstNumber);
         }
 
         [Given(@"the second number is (.*)")]
-        public void GivenTheSecondNumberIs(int secondNumber)
+        public void GivenTheSecondNumberIs(double secondNumber)
         {
             _scenarioContext.Add("SecondNumber", secondNumber);
         }
@@ -33,39 +33,39 @@ namespace bdd.workshop.calculator.tests.tdd.steps
         [When(@"the two numbers are added")]
         public void WhenTheTwoNumbersAreAdded()
         {
-            var firstNumber = _scenarioContext.Get<int>("FirstNumber");
-            var secondNumber = _scenarioContext.Get<int>("SecondNumber");
+            var firstNumber = _scenarioContext.Get<double>("FirstNumber");
+            var secondNumber = _scenarioContext.Get<double>("SecondNumber");
             _scenarioContext.Add("Result", Operator.Add(firstNumber, secondNumber));
         }
 
         [Then(@"the result should be (.*)")]
         [Then(@"the result shall be (.*)")]
-        public void ThenTheResultShouldBe(int result)
+        public void ThenTheResultShouldBe(double result)
         {
-            Assert.True(result == _scenarioContext.Get<int>("Result"));
+            Assert.True(result == _scenarioContext.Get<double>("Result"));
         }
 
         [When(@"I substract first number to second number")]
         public void WhenISubstractFirstNumberToSecondNumber()
         {
-            var firstNumber = _scenarioContext.Get<int>("FirstNumber");
-            var secondNumber = _scenarioContext.Get<int>("SecondNumber");
+            var firstNumber = _scenarioContext.Get<double>("FirstNumber");
+            var secondNumber = _scenarioContext.Get<double>("SecondNumber");
             _scenarioContext.Add("Result", Operator.Substract(firstNumber, secondNumber));
         }
 
         [When(@"I multiply both numbers")]
         public void WhenIMultiplyBothNumbers()
         {
-            var firstNumber = _scenarioContext.Get<int>("FirstNumber");
-            var secondNumber = _scenarioContext.Get<int>("SecondNumber");
+            var firstNumber = _scenarioContext.Get<double>("FirstNumber");
+            var secondNumber = _scenarioContext.Get<double>("SecondNumber");
             _scenarioContext.Add("Result", Operator.Multiply(firstNumber, secondNumber));
         }
 
         [When(@"I divide first number by second number")]
         public void WhenIDivideFirstNumberBySecondNumber()
         {
-            var firstNumber = _scenarioContext.Get<int>("FirstNumber");
-            var secondNumber = _scenarioContext.Get<int>("SecondNumber");
+            var firstNumber = _scenarioContext.Get<double>("FirstNumber");
+            var secondNumber = _scenarioContext.Get<double>("SecondNumber");
             _scenarioContext.Add("Result", Operator.Divide(firstNumber, secondNumber));
         }
